[#118] move emotion cache file
